refactor(integrations): use client-side rendering package
