+++ .changeset/fair-mangos-tickle.md
@@ -1,0 +1,11 @@
+---
+'@scalar/astro': patch
+'@scalar/express-api-reference': patch
+'@scalar/fastify-api-reference': patch
+'@scalar/hono-api-reference': patch
+'@scalar/nestjs-api-reference': patch
+'@scalar/nextjs-api-reference': patch
+'@scalar/sveltekit': patch
+---
+
+refactor: migrate integrations to client-side rendering package
