+++ integrations/astro/src/ScalarComponent.astro
@@ -1,5 +1,6 @@
 ---
-import { getHtmlDocument, type HtmlRenderingConfiguration } from '@scalar/core/libs/html-rendering'
+import { renderApiReference } from '@scalar/client-side-rendering'
+import type { HtmlRenderingConfiguration } from '@scalar/types/api-reference'
 
 export interface Props {
   configuration: Partial<HtmlRenderingConfiguration>
@@ -15,9 +16,15 @@ const DEFAULT_CONFIGURATION: Partial<HtmlRenderingConfiguration> = {
 
 const { configuration } = Astro.props
 
-const html = getHtmlDocument({
+const finalConfiguration = {
   ...DEFAULT_CONFIGURATION,
   ...configuration,
+}
+const { cdn, pageTitle, ...config } = finalConfiguration
+const html = renderApiReference({
+  config,
+  cdn,
+  pageTitle,
 })
 ---
