+++ integrations/express/package.json
@@ -54,7 +54,8 @@
     "documentation": "https://scalar.com/products/api-references/integrations/express"
   },
   "dependencies": {
-    "@scalar/core": "workspace:*"
+    "@scalar/client-side-rendering": "workspace:*",
+    "@scalar/types": "workspace:*"
   },
   "devDependencies": {
     "@types/express": "catalog:*",
