+++ integrations/fastify/package.json
@@ -61,9 +61,10 @@
     "documentation": "https://scalar.com/products/api-references/integrations/fastify"
   },
   "dependencies": {
-    "@scalar/core": "workspace:*",
+    "@scalar/client-side-rendering": "workspace:*",
     "@scalar/openapi-parser": "workspace:*",
     "@scalar/openapi-types": "workspace:*",
+    "@scalar/types": "workspace:*",
     "fastify-plugin": "catalog:*",
     "github-slugger": "catalog:*"
   },
