+++ integrations/fastify/src/fastifyApiReference.ts
@@ -1,5 +1,5 @@
 /// <reference types="@fastify/swagger" />
-import { getHtmlDocument } from '@scalar/core/libs/html-rendering'
+import { renderApiReference } from '@scalar/client-side-rendering'
 import { normalize, toJson, toYaml } from '@scalar/openapi-parser'
 import type { OpenAPI } from '@scalar/openapi-types'
 import type { FastifyBaseLogger, FastifyTypeProviderDefault, RawServerDefault } from 'fastify'
@@ -226,13 +226,17 @@ const fastifyApiReference = fp<
         }
 
         // Respond with the HTML document
-        return reply.header('Content-Type', 'text/html; charset=utf-8').send(
-          getHtmlDocument({
-            // We're using the bundled JS here by default, but the user can pass a CDN URL.
-            cdn: RELATIVE_JAVASCRIPT_PATH,
-            ...configuration,
-          }),
-        )
+        const { cdn, pageTitle, ...config } = configuration
+        return reply
+          .header('Content-Type', 'text/html; charset=utf-8')
+          .send(
+            renderApiReference({
+              config,
+              // We're using the bundled JS here by default, but the user can pass a CDN URL.
+              cdn: cdn ?? RELATIVE_JAVASCRIPT_PATH,
+              pageTitle,
+            }),
+          )
       },
     })
