+++ integrations/hono/package.json
@@ -52,7 +52,8 @@
     "documentation": "https://scalar.com/products/api-references/integrations/hono"
   },
   "dependencies": {
-    "@scalar/core": "workspace:*"
+    "@scalar/client-side-rendering": "workspace:*",
+    "@scalar/types": "workspace:*"
   },
   "devDependencies": {
     "@hono/node-server": "catalog:*",
