+++ integrations/nestjs/package.json
@@ -64,7 +64,8 @@
     "documentation": "https://scalar.com/products/api-references/integrations/nestjs"
   },
   "dependencies": {
-    "@scalar/core": "workspace:*"
+    "@scalar/client-side-rendering": "workspace:*",
+    "@scalar/types": "workspace:*"
   },
   "devDependencies": {
     "@nestjs/common": "catalog:*",
