+++ integrations/sveltekit/package.json
@@ -64,7 +64,8 @@
   ],
   "svelte": "./dist/index.js",
   "dependencies": {
-    "@scalar/core": "workspace:*"
+    "@scalar/client-side-rendering": "workspace:*",
+    "@scalar/types": "workspace:*"
   },
   "devDependencies": {
     "@sveltejs/kit": "^2.57.1",
