+++ integrations/sveltekit/src/types.ts
@@ -1,4 +1,4 @@
-import type { HtmlRenderingConfiguration } from '@scalar/core/libs/html-rendering'
+import type { HtmlRenderingConfiguration } from '@scalar/types/api-reference'
 
 /**
  * The configuration for the Scalar API Reference for SvelteKit
